The 9x9 Solver
Hey! This is my first project I've made and it's super smooth. Essentially, you can use this to solve Sudoku's using a back fill method. What this means is, the solver will work line by line, inputting integers from 1-9, ensuring that there are no repeat values in the row, column, or square, for that variable. Once it is finished, it will print out the board. I've tested this 10 times and it has it boosted my time to completion dramatically. I hope you enjoy!